Improved data management for Player.js
There seems to be confusion amongst all of us on how we're using each
others code. There's a lot of areas in which access hierarchy is broken,
extra parameters being given that aren't part of the functions'
definitions, etc... etc...
P.S. For now I commented out the lines that add the civilians. There's
definitely something wrong with it that is making the game untestable,
even on my personal pc.