I FOUND THE BUG WITHIN CIVILIAN!!!
Civilian uses a function called inArray that's located in util.js. The
issue is that the second parameter it takes is an array, and the array
being given is ridiculously huge. Like there's an error on how large
it's getting huge. Idk if the bug is within inArray, or if the array
called 'visited' is being given too many indexes, but the problem's
there.