Added a scrollable camera
Now our level size isn't constrained to the size of the screen being
played on. The camera will scroll with the player. Dante, I made sure to
do it in such a way that it wouldn't mess with your AI and grids, but
tomorrow lets get together to make sure of it. It's an easy process to
remove the scrolling feature if it doesn't work